Removed the computation of tiles from the GeometryAttribute class
Mixen of Clojure (tiles/nl) and Java (GeometryAttribute) in one class (GeometryAttribute) gives dependency-problems.
So the computation of tiles is extracted from the class GeometryAttribute. The computation have to done outside of
this class. The result (set) of this computation have to be added to GeometryAttribute with the method setTiles.

Ref. PDOK-3512