Use GeometryAttribute (Java) to hold data for geometries
* Moved from the geometry-map (Clojure) to GeometryAttribute class (Java).
To make geometry-information serializable (and exchangable) to other (Java-)clients it was needed to use a Java-class for the geometry-information.

* Removed the computation of tiles from the GeometryAttribute class

Mixen of Clojure (tiles/nl) and Java (GeometryAttribute) in one class (GeometryAttribute) gives dependency-problems.
So the computation of tiles is extracted from the class GeometryAttribute. The computation have to done outside of
this class. The result (set) of this computation have to be added to GeometryAttribute with the method setTiles.

Ref. PDOK-3512

* Add GeometryAttribute to TransitHandlers

[PDOK-3512]

* Rewrite the tests from TilesHelperTest in Clojure

Moved some tests to the Clojure section. This to document
how to use TilesHelper in Clojure functions.

[PDOK-3512]

* Enabled (again) failing LocalDateTimeZeroTest in TransithandlersTest

The failing LocalDateTimeZeroTest was disabled. To write and test
functionality of GeometeryAttribute the test was disabled, so
functionality could be build. In next commit it is needed that test/
code wil be fixed.

[PDOK-3512]

* Fixed LocalDateTimeZeroTest

Something has be gone wrong with merging master- and feature-branche.
Comparing and using code form master-branche has solved the failing test.

[PDOK-3512]

* Changed tiles from Collection to Set

GeometryAttribute has to be serialized with the transit library/format
this format doesn't support Collection but Set.

[PDOK-3512]

* Use the stored tiles of GeometryAttribute (if present)

[PDOK-3512]

* No NullPointerExecption when GeometryAttribute is nil/null

[PDOK-3512]

* Tiles in datastructure GeometryAttribute these will be transit-serialized

[PDOK-3512]

* Fix test TransitHandler GeometryAttribute

[PDOK-3512]

* Correct typo "dispatcher"

[PDOK-3512]

* Cleanup imports

[PDOK-3512]